More final code touchups
